refractor: created a separate test for inherited properties and methods for the Animal class
