+++ src/Animals/Bird.cs
@@ -8,6 +8,5 @@ namespace Animals.Bird
 {
     public abstract class Bird:Animal
     {      
-       public Bird(int id) : base(id) { }
     }
 }
