+++ src/Animals/Bison.cs
@@ -10,7 +10,6 @@ namespace Animals.Mammal
     {
         public override string[] FavoriteFood { get; } = new string[] { "grass", "vegetable" };
         public override int RequiredSpaceSqFt { get; } = 1000;
-        public Bison(int id) : base(id) { }
         public override bool IsFriendlyWithAnimal(Animal animal)
         {
             string FriendlyAnimalList = "Elephant, Bison";
