+++ src/Animals/Elephant.cs
@@ -10,7 +10,6 @@ namespace Animals.Mammal
     {
         public override string[] FavoriteFood { get; } = new string[] { "grass", "vegetable", "fruit" };
         public override int RequiredSpaceSqFt { get; } = 1000;
-        public Elephant(int id) : base(id) { }
         public override bool IsFriendlyWithAnimal(Animal animal)
         {
             string FriendlyAnimalList = "Bison, Parrot, Turtle";
