+++ src/Animals/Lion.cs
@@ -10,7 +10,6 @@ namespace Animals.Mammal
     {
         public override string[] FavoriteFood { get; } = new string[] { "meet" };
         public override int RequiredSpaceSqFt { get; } = 1000;
-        public Lion(int id) : base(id) { }
         public override bool IsFriendlyWithAnimal(Animal animal)
         {
             string FriendlyAnimalList = "Lion";
