+++ src/Animals/Parrot.cs
@@ -10,7 +10,6 @@ namespace Animals.Bird
     {
         public override string[] FavoriteFood { get; } = new string[] { "fruit", "vegetables" };
         public override int RequiredSpaceSqFt { get; } = 5;
-        public Parrot(int id) : base(id) { }
         public override bool IsFriendlyWithAnimal(Animal animal)
         {
             string FriendlyAnimalList = "Parrot, Bison, Elephant, Turtle";
