+++ src/Animals/Penguin.cs
@@ -4,8 +4,6 @@
     {
         public override string[] FavoriteFood { get; } = new string[] {"salmon", "herring" };
         public override int RequiredSpaceSqFt { get; } = 10;
-        public Penguin(int id) : base(id) { }
-
         public override bool IsFriendlyWithAnimal(Animal animal)
         {
             string FriendlyAnimalList = "Penguin";
