+++ src/Enclousers/Enclouser.cs
@@ -31,13 +31,7 @@ namespace Enclousers
                     throw new NotFriendlyAnimalException();
                 }
             }
-            var random = new Random();
-            int randomId = random.Next(0, int.MaxValue);
-            while (Animals.Any(note => note.ID == randomId))
-            {
-                randomId = random.Next(0, int.MaxValue);
-            }
-            animal.ID = randomId;
+            
             Animals.Add(animal);
             SqureFeet -= animal.RequiredSpaceSqFt;
         }
