+++ tests/AnimalsTests/ElephantTest.cs
@@ -10,42 +10,6 @@ namespace AnimalsTests
     public class ElephantTest
     {
         [Fact]
-        public void ShouldBeHealed()
-        {
-            var elephant = new Elephant();
-            Antibiotics antibiotics = new();
-            elephant.Heal(antibiotics);
-            Assert.False(elephant.IsSick);
-            elephant = new Elephant();
-            AntiDepression antiDepression = new();
-            elephant.Heal(antiDepression);
-            Assert.False(elephant.IsSick);
-            elephant = new Elephant();
-            AntiInflammatory antiInflammatory = new();
-            elephant.Heal(antiInflammatory);
-            Assert.False(elephant.IsSick);
-        }
-        [Theory]
-        [InlineData(0, 23)]
-        [InlineData(5, 20)]
-        public void ShouldAddSchedule(int value1, int value2)
-        {
-            var elephant = new Elephant();
-            List<int> moreFeedTimes = new() { value1, value2 };
-            elephant.AddSchedule(moreFeedTimes);
-            Assert.Equal(value1, elephant.FeedSchedule[0]);
-            Assert.Equal(value2, elephant.FeedSchedule[elephant.FeedSchedule.Count - 1]);
-        }
-        [Theory]
-        [InlineData(547)]
-        [InlineData(6698705)]
-        public void ShouldBeCorrectID(int value)
-        {
-            var elephant = new Elephant();
-            elephant.ID = value;
-            Assert.Equal(value, elephant.ID);
-        }
-        [Fact]
         public void ShouldBeFriendly()
         {
             var elephant = new Elephant();
