+++ tests/AnimalsTests/PenguinTest.cs
@@ -12,42 +12,6 @@ namespace AnimalsTests
     public class PenguinTest
     {
         [Fact]
-        public void ShouldBeHealed()
-        {
-            Penguin penguin = new Penguin();
-            Antibiotics antibiotics = new();
-            penguin.Heal(antibiotics);
-            Assert.False(penguin.IsSick);
-            penguin = new Penguin();
-            AntiDepression antiDepression = new();
-            penguin.Heal(antiDepression);
-            Assert.False(penguin.IsSick);
-            penguin = new Penguin();
-            AntiInflammatory antiInflammatory = new();
-            penguin.Heal(antiInflammatory);
-            Assert.False(penguin.IsSick);
-        }
-        [Theory]
-        [InlineData(0, 23)]
-        [InlineData(5, 20)]
-        public void ShouldAddSchedule(int value1, int value2)
-        {
-            Penguin penguin = new Penguin();
-            List<int> moreFeedTimes = new() { value1, value2 };
-            penguin.AddSchedule(moreFeedTimes);
-            Assert.Equal(value1, penguin.FeedSchedule[0]);
-            Assert.Equal(value2, penguin.FeedSchedule[penguin.FeedSchedule.Count - 1]);
-        }
-        [Theory]
-        [InlineData(547)]
-        [InlineData(6698705)]
-        public void ShouldBeCorrectID(int value)
-        {
-            Penguin penguin = new Penguin();
-            penguin.ID = value;
-            Assert.Equal(value, penguin.ID);
-        }
-        [Fact]
         public void ShouldBeFriendly()
         {
             var penguin = new Penguin();
