+++ tests/AnimalsTests/SnakeTest.cs
@@ -12,42 +12,6 @@ namespace AnimalsTests
    public class SnakeTest
     {
        [Fact]
-       public void ShouldBeHealed()
-       {
-           var snake = new Snake();
-           var antibiotics = new Antibiotics();
-           snake.Heal(antibiotics);
-           Assert.False(snake.IsSick);
-           snake = new Snake();
-           var antiDepression = new AntiDepression();
-           snake.Heal(antiDepression);
-           Assert.False(snake.IsSick);
-           snake = new Snake();
-           var antiInflammatory = new AntiInflammatory();
-           snake.Heal(antiInflammatory);
-           Assert.False(snake.IsSick);
-       }
-       [Theory]
-       [InlineData(0, 23)]
-       [InlineData(5, 20)]
-       public void ShouldAddSchedule(int value1, int value2)
-       {
-            var snake = new Snake();
-            List<int> moreFeedTimes = new() { value1, value2 };
-            snake.AddSchedule(moreFeedTimes);
-           Assert.Equal(value1, snake.FeedSchedule[0]);
-           Assert.Equal(value2, snake.FeedSchedule[snake.FeedSchedule.Count - 1]);
-       }
-       [Theory]
-       [InlineData(547)]
-       [InlineData(6698705)]
-       public void ShouldBeCorrectID(int value)
-       {
-           var snake = new Snake();
-           snake.ID = value;
-           Assert.Equal(value, snake.ID);
-       }
-       [Fact]
        public void ShouldBeFriendly()
        {
             var snake = new Snake();
